+++ mayan/apps/databases/literals.py
@@ -1,0 +1,3 @@
+IMPORT_ERROR_EXCLUSION_TEXTS = (
+    'doesn\'t look like a module path', 'No module named'
+)