+++ mayan/apps/documents/models/favorite_document_models.py
@@ -5,7 +5,7 @@ from django.db import models
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.model_mixins import ExtraDataModelMixin
+from mayan.apps.databases.model_mixins import ExtraDataModelMixin
 
 from ..managers import FavoriteDocumentManager