+++ mayan/apps/web_links/models.py
@@ -3,7 +3,7 @@ from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 
-from mayan.apps.common.model_mixins import ExtraDataModelMixin
+from mayan.apps.databases.model_mixins import ExtraDataModelMixin
 from mayan.apps.documents.models import DocumentType
 from mayan.apps.events.classes import EventManagerSave
 from mayan.apps.events.decorators import method_event